Create board model and dummy data
