--- script/board-dummy.js
@@ -1,200 +1,0 @@
-const boardData = [
-	{
-		name: 'Zathin',
-		type: 'team'
-	},
-	{
-		name: 'Stim',
-		type: 'team'
-	},
-	{
-		name: 'Redhold',
-		type: 'team'
-	},
-	{
-		name: '',
-		type: 'personal'
-	},
-	{
-		name: 'Overhold',
-		type: 'personal'
-	},
-	{
-		name: 'Prodder',
-		type: 'personal'
-	},
-	{
-		name: 'Bytecard',
-		type: 'personal'
-	},
-	{
-		name: 'Regrant',
-		type: 'team'
-	},
-	{
-		name: 'Zaam-Dox',
-		type: 'team'
-	},
-	{
-		name: 'Opela',
-		type: 'personal'
-	},
-	{
-		name: '',
-		type: 'personal'
-	},
-	{
-		name: 'Lotstring',
-		type: 'team'
-	},
-	{
-		name: 'Greenlam',
-		type: 'personal'
-	},
-	{
-		name: 'Zoolab',
-		type: 'personal'
-	},
-	{
-		name: 'Lotlux',
-		type: 'team'
-	},
-	{
-		name: 'Quo Lux',
-		type: 'team'
-	},
-	{
-		name: 'Redhold',
-		type: 'team'
-	},
-	{
-		name: 'Sub-Ex',
-		type: 'team'
-	},
-	{
-		name: 'Flowdesk',
-		type: 'personal'
-	},
-	{
-		name: 'Span',
-		type: 'team'
-	},
-	{
-		name: 'Quo Lux',
-		type: 'personal'
-	},
-	{
-		name: 'Y-Solowarm',
-		type: 'team'
-	},
-	{
-		name: 'Tampflex',
-		type: 'team'
-	},
-	{
-		name: '',
-		type: 'personal'
-	},
-	{
-		name: 'Pannier',
-		type: 'personal'
-	},
-	{
-		name: 'Hatity',
-		type: 'personal'
-	},
-	{
-		name: 'Sonair',
-		type: 'team'
-	},
-	{
-		name: 'Mat Lam Tam',
-		type: 'team'
-	},
-	{
-		name: 'Tempsoft',
-		type: 'team'
-	},
-	{
-		name: 'Subin',
-		type: 'team'
-	},
-	{
-		name: 'Zathin',
-		type: 'personal'
-	},
-	{
-		name: 'Sub-Ex',
-		type: 'team'
-	},
-	{
-		name: 'Veribet',
-		type: 'team'
-	},
-	{
-		name: 'Tresom',
-		type: 'team'
-	},
-	{
-		name: 'Bigtax',
-		type: 'personal'
-	},
-	{
-		name: 'Tampflex',
-		type: 'personal'
-	},
-	{
-		name: 'Konklux',
-		type: 'personal'
-	},
-	{
-		name: 'Veribet',
-		type: 'team'
-	},
-	{
-		name: 'Latlux',
-		type: 'team'
-	},
-	{
-		name: 'Biodex',
-		type: 'team'
-	},
-	{
-		name: 'Cookley',
-		type: 'team'
-	},
-	{
-		name: 'It',
-		type: 'personal'
-	},
-	{
-		name: 'Alphazap',
-		type: 'personal'
-	},
-	{
-		name: 'Voltsillam',
-		type: 'team'
-	},
-	{
-		name: 'Gembucket',
-		type: 'team'
-	},
-	{
-		name: 'Viva',
-		type: 'team'
-	},
-	{
-		name: 'Stronghold',
-		type: 'team'
-	},
-	{
-		name: 'Aerified',
-		type: 'team'
-	},
-	{
-		name: '',
-		type: 'personal'
-	}
-];
-
-module.exports = boardData;
