+++ server/db/models/board.js
@@ -1,20 +1,20 @@
-const sequelize = require('sequelize');
-const db = require('../db');
+const Sequelize = require('sequelize')
+const db = require('../db')
 
 const Board = db.define('board', {
-	name: {
-		type: Sequelize.STRING,
-		defaultValue: 'Board',
-		validate: {
-			len: [ 1, 150 ]
-		}
-	},
-	type: {
-		type: Sequelize.ENUM({
-			values: [ 'personal', 'team' ]
-		}),
-		allowNull: false
-	}
-});
+  name: {
+    type: Sequelize.STRING,
+    defaultValue: 'Board',
+    validate: {
+      len: [1, 150]
+    }
+  },
+  type: {
+    type: Sequelize.ENUM({
+      values: ['personal', 'team']
+    }),
+    allowNull: false
+  }
+})
 
-module.exports = Board;
+module.exports = Board
