+++ server/db/models/index.js
@@ -1,5 +1,5 @@
 const User = require('./user')
-
+const Board = require('./board')
 /**
  * If we had any associations to make, this would be a great place to put them!
  * ex. if we had another model called BlogPost, we might say:
@@ -14,5 +14,6 @@ const User = require('./user')
  * instead of: const User = require('../db/models/user')
  */
 module.exports = {
-  User
+  User,
+  Board
 }
